third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-promise-reject-and-handle-expected.txt

CONSOLE MESSAGE: line 14: onunhandledrejection1
CONSOLE MESSAGE: line 21: onrejectionhandled1
CONSOLE MESSAGE: line 29: onunhandledrejection2
CONSOLE ERROR: line 27: Uncaught (in promise) Error: e
CONSOLE MESSAGE: line 35: onrejectionhandled2
Tests that evt.preventDefault() in window.onunhandledrejection suppresses console output.

----console messages start----
console-promise-reje…-and-handle.html:14 onunhandledrejection1
console-promise-reje…-and-handle.html:21 onrejectionhandled1
console-promise-reje…-and-handle.html:29 onunhandledrejection2
console-promise-reje…-and-handle.html:27 Uncaught (in promise) Error: e
    at setupPromise2 (console-promise-reje…-and-handle.html:27)
console-promise-reje…-and-handle.html:35 onrejectionhandled2
----console messages end----
